long-image refactor as LongImageComponent
