+++ haoshiyou/src/pages/listings-tab/long-image.comp.html
@@ -1,0 +1,4 @@
+<button ion-button color="primary" clear item-left (click)="generateLongImage()">
+    下载长图片 <ion-icon name="ios-long-image"></ion-icon>
+</button>
+<a id="downloadLink" hidden></a>
